代码架构大调整

+++ main.py
@@ -4,17 +4,19 @@ from torch import embedding, nn
 from utils import sc_normalization, mask_data, construct_graph, \
     read_data_label_h5, read_similarity_mat_h5, \
     cpm_classify, z_score_normalization, show_cluster, \
-    concat_views, batch_mixing_entropy, runPCA, runUMAP
+    concat_views, batch_mixing_entropy, runPCA, runUMAP, encode_label
 from model import scGNN, CPMNets, Classifier
 import numpy as np
 from sklearn.metrics import silhouette_score, adjusted_rand_score
 import wandb
-from sklearn import preprocessing
+import pickle
+
+
 # seq_well_smart 只有五类!!!
 # drop_seq_10x_v3有8类
 
 # 训练scGNN，得到每个Pathway的embedding
-def train_scGNN(model, n_epochs, G_data, optimizer,
+def train_scGCN(model, G_data, optimizer,
                 index_pair, masking_idx, norm_data, loss_title):
     '''
     :param model: 待训练的模型
@@ -27,7 +29,7 @@ def train_scGNN(model, n_epochs, G_data, optimizer,
     :return:
     '''
     model = model.to(device)
-    for epoch in range(n_epochs):
+    for epoch in range(parameter_config['epoch_GCN']):
         model.train()
         optimizer.zero_grad()
         pred = model(G_data.to(device))
@@ -47,102 +49,104 @@ def train_scGNN(model, n_epochs, G_data, optimizer,
     return model
 
 
-def train_cpm_net(ref_data_embeddings: torch.Tensor,
-                  ref_label: torch.Tensor,
-                  query_data_embeddings: torch.Tensor,
-                  query_label: torch.Tensor,
-                  ref_view_num: int,
-                  ref_view_feat_len: list,
-                  config: dict):
-    train_len = ref_data_embeddings.shape[0]
-    test_len = query_data_embeddings.shape[0]
-    # lsd_dim 作为超参数可调
-    model = CPMNets(ref_view_num, train_len, test_len, ref_view_feat_len, config['ref_class_num'], config['lsd_dim'],
-                    config)
-
-    # 开始训练
-    model.train_ref_h(ref_data_embeddings, ref_label, config['epoch_CPM_train'], config['CPM_lr'])
-
-    # 对test_h进行adjust（按照论文的想法，保证consistency）
-    model.train_query_h(query_data_embeddings, ref_label, query_label, config['epoch_CPM_test'])
-
-    ref_h = model.get_h_train().detach().cpu().numpy()
-    query_h = model.get_h_test().detach().cpu().numpy()
-    
-    return model, ref_h, query_h
-
-def self_supervised_train(data, data_name, data_path, config):
+def self_supervised_train(data):
     # 可以试试调整这个mask比例来调参
-    masked_prob = min(len(data.nonzero()[0]) / (data.shape[0] * data.shape[1]), config["mask_rate"])
+    masked_prob = min(len(data.nonzero()[0]) / (data.shape[0] * data.shape[1]), parameter_config["mask_rate"])
     masked_data, index_pair, masking_idx = mask_data(data, masked_prob)
 
-    sm_arr = [read_similarity_mat_h5(data_path, "sm_" + data_name + "_" + str(i + 1)) for i in
-                  range(4)]
-    graphs = [construct_graph(masked_data, sm_arr[i], config['k']) for i in range(len(sm_arr))]
+    sm_arr = [read_similarity_mat_h5(data_config['data_path'], "ref/sm_" + str(i + 1)) for i in
+              range(4)]
+
+    graphs = [construct_graph(masked_data, sm_arr[i], parameter_config['k']) for i in range(len(sm_arr))]
 
     embeddings_arr = []
-    GNN_models = []
+    gcn_models = []
 
     # 训练ref data in scGNN
     for i in range(len(graphs)):
-        model = scGNN(graphs[i], config['middle_out'])
-        optimizer = torch.optim.Adam(model.parameters(), lr=config['GNN_lr'])
-        model = train_scGNN(model, config['epoch_GCN'], graphs[i], optimizer, index_pair, masking_idx,
-                            data, data_name +': GNN: view' + str(i + 1) + ' loss')
+        model = scGNN(graphs[i], parameter_config['middle_out'])
+        optimizer = torch.optim.Adam(model.parameters())
+        model = train_scGCN(model, graphs[i], optimizer, index_pair, masking_idx,
+                            data, 'GCN: view' + str(i + 1) + ' loss')
 
         # 利用未mask的矩阵，构造图，丢入训练好的model，得到中间层embedding
         # embedding = model.get_embedding(Graph(scDataNorm, similarity_matrix_arr[i]))
         # 还是用mask好的数据得到Embedding比较符合自监督的逻辑
         embedding = model.get_embedding(graphs[i])
         embeddings_arr.append(embedding.detach().cpu().numpy())
-        GNN_models.append(model)
-    return embeddings_arr, GNN_models
+        gcn_models.append(model)
+    return embeddings_arr, gcn_models
 
 
-def transfer_train(data_config: dict,
-                   config: dict):
+def train_query(gcn_models, cpm_model, query_data):
+    # 数据预处理
+    query_norm_data = sc_normalization(query_data)
+
+    # 构造Query data的Graph
+    query_sm_arr = [read_similarity_mat_h5(data_config['data_path'], "query/sm_" + str(i + 1)) for i in
+                    range(4)]
+    query_graphs = [construct_graph(query_norm_data, query_sm_arr[i], parameter_config['k'])
+                    for i in range(len(query_sm_arr))]
+
+    # 获得Embedding
+    query_views = []
+    for i in range(len(gcn_models)):
+        query_views.append(gcn_models[i].get_embedding(query_graphs[i]).detach().cpu().numpy())
+
+    query_data_embeddings_tensor = torch.from_numpy(z_score_normalization(concat_views(query_views))).float().to(device)
+
+    # query_data_embeddings_tensor = torch.from_numpy(concat_views(query_views)).float().to(device)
+    # query_label_tensor = torch.from_numpy(query_labels).view(1, query_labels.shape[0]).long().to(device)
+
+    query_h = cpm_model.train_query_h(query_data_embeddings_tensor, parameter_config['epoch_CPM_test'])
+    return query_h
+
+
+def transfer_labels():
+    if parameter_config['model_exist']:
+        # 如果模型存在, 加载已存在的模型，然后对query进行重构
+        gcn_models, cpm_model = load_models()
+        query_data, query_label = read_data_label_h5(data_config['data_path'], data_config['query_key'])
+        query_data = query_data.astype(np.float64)
+
+        ref_h = cpm_model.get_h_train()
+        ref_label = cpm_model.get_ref_labels()
+
+        query_h = train_query(gcn_models, cpm_model, query_data)
+
+        pred = cpm_classify(ref_h, query_h, ref_label)
+        acc = (pred == query_label).sum()
+        acc = acc / pred.shape[0]
+
+        ret = {
+            'acc': acc,
+            'ref_h': ref_h,
+            'query_h': query_h,
+            'ref_label': ref_label,
+            'query_raw_data': query_data,
+            'query_label': query_label,
+            'pred': pred,
+        }
+        return ret
+
+
     '''
-    :param data_path: 表达矩阵的路径
-    :param label_path: 标签路径
-    :param sm_path: 相似矩阵的路径
-    :param config: 相关参数的配置
-    :return:
+        数据准备
     '''
-    data_path = os.path.join(data_config['data_path'], 'data.h5')
+    # data_path = os.path.join(data_config['data_path'], 'data.h5')
 
-    ref_data, ref_label = read_data_label_h5(data_path, data_config['ref_name'])
-    
+    ref_data, ref_label = read_data_label_h5(data_config['data_path'], "ref")
     ref_data = ref_data.astype(np.float64)
-    ref_enc = preprocessing.LabelEncoder()
-    # label从0开始
-    ref_label = (ref_enc.fit_transform(ref_label)).astype(np.int64)
+
+    query_data, query_label = read_data_label_h5(data_config['data_path'], data_config['query_key'])
+    query_data = query_data.astype(np.float64)
+
+    # 将ref和query data进行编码
+    ref_label, query_label, enc = encode_label(ref_label, query_label)
 
     # 数据预处理
     ref_norm_data = sc_normalization(ref_data)
-    ref_views, ref_gcn_models = self_supervised_train(ref_norm_data, data_config['ref_name'], data_path, config)
-    # 可以试试调整这个mask比例来调参
-    # masked_prob = min(len(ref_norm_data.nonzero()[0]) / (ref_norm_data.shape[0] * ref_norm_data.shape[1]), 0.2)
-    # masked_ref_data, index_pair, masking_idx = mask_data(ref_norm_data, masked_prob)
-    #
-    # ref_sm_arr = [read_similarity_mat_h5(data_path, "SM_"+data_config["ref_name"]+"_"+str(i+1)) for i in range(4)]
-    # ref_graphs = [construct_graph(masked_ref_data, ref_sm_arr[i], config['k']) for i in range(len(ref_sm_arr))]
-    #
-    # ref_views = []
-    # GNN_models = []
-    #
-    # # 训练ref data in scGNN
-    # for i in range(len(ref_graphs)):
-    #     model = scGNN(ref_graphs[i], config['middle_out'])
-    #     optimizer = torch.optim.Adam(model.parameters(), lr=config['GNN_lr'])
-    #     model = train_scGNN(model, config['epoch_GCN'], ref_graphs[i], optimizer, index_pair, masking_idx,
-    #                         ref_norm_data, 'GNN: view' + str(i + 1) + ' loss')
-    #
-    #     # 利用未mask的矩阵，构造图，丢入训练好的model，得到中间层embedding
-    #     # embedding = model.get_embedding(Graph(scDataNorm, similarity_matrix_arr[i]))
-    #     # 还是用mask好的数据得到Embedding比较符合自监督的逻辑
-    #     embedding = model.get_embedding(ref_graphs[i])
-    #     ref_views.append(embedding.detach().cpu().numpy())
-    #     GNN_models.append(model)
+    ref_views, ref_gcn_models = self_supervised_train(ref_norm_data)
 
     '''
         Reference Data Embeddings  
@@ -154,60 +158,36 @@ def transfer_train(data_config: dict,
     for i in range(ref_view_num):
         ref_view_feat_len.append(ref_views[i].shape[1])
 
-    # 把所有的view连接在一起
+    # 把所有的view连接在一起, 并做一个归一化
     ref_data_embeddings_tensor = torch.from_numpy(z_score_normalization(concat_views(ref_views))).float().to(device)
-    # ref_data_embeddings_tensor = torch.from_numpy(concat_views(ref_views)).float().to(device)
     ref_label_tensor = torch.from_numpy(ref_label).view(1, ref_label.shape[0]).long().to(device)
 
     '''
-        Query data
+        对ref data进行训练
     '''
-    query_data, query_label = read_data_label_h5(data_path, data_config['query_name'])
-    query_data = query_data.astype(np.float64)
-    query_enc = preprocessing.LabelEncoder()
-    query_label = (query_enc.fit_transform(query_label)).astype(np.int64)
+    cpm_model = CPMNets(ref_view_num,
+                        ref_data.shape[0],
+                        ref_view_feat_len,
+                        ref_label,
+                        parameter_config)
 
-    # 数据预处理
-    query_norm_data = sc_normalization(query_data)
-    # query_norm_data = preprocessing.minmax_scale(query_norm_data, axis=0)
+    # 开始训练
+    cpm_model.train_ref_h(ref_data_embeddings_tensor, ref_label_tensor)
 
-    # 构造Query data的Graph
-    query_sm_arr = [read_similarity_mat_h5(data_path, "sm_"+data_config["query_name"]+"_"+str(i+1)) for i in range(4)]
-    query_graphs = [construct_graph(query_norm_data, query_sm_arr[i], config['k']) for i in range(len(query_sm_arr))]
-
-    # 获得Embedding
-    # query_views, query_gcn_models = self_supervised_train(query_norm_data, data_config['query_name'], data_path, config)
-    query_views = []
-    for i in range(len(ref_gcn_models)):
-        query_views.append(ref_gcn_models[i].get_embedding(query_graphs[i]).detach().cpu().numpy())
-
-    query_data_embeddings_tensor = torch.from_numpy(z_score_normalization(concat_views(query_views))).float().to(device)
-
-    # query_data_embeddings_tensor = torch.from_numpy(concat_views(query_views)).float().to(device)
-    query_label_tensor = torch.from_numpy(query_label).view(1, query_label.shape[0]).long().to(device)
-
-    '''
-        CPM-Net
-    '''
-    cpm_model, ref_h, query_h = train_cpm_net(ref_data_embeddings_tensor,
-                                              ref_label_tensor,
-                                              query_data_embeddings_tensor,
-                                              query_label_tensor,
-                                              ref_view_num,
-                                              ref_view_feat_len,
-                                              config)
-
+    # 得到最后的embeddings (ref和query)
+    ref_h = cpm_model.get_h_train()
+    query_h = train_query(ref_gcn_models, cpm_model, query_data)
 
     pred = cpm_classify(ref_h, query_h, ref_label)
     acc = (pred == query_label).sum()
     acc = acc / pred.shape[0]
 
     # 还原label
-    ref_label = ref_enc.inverse_transform(ref_label)
+    ref_label = enc.inverse_transform(ref_label)
 
-    query_label = query_enc.inverse_transform(query_label)
+    query_label = enc.inverse_transform(query_label)
 
-    pred = query_enc.inverse_transform(pred)
+    pred = enc.inverse_transform(pred)
 
     ret = {
         'acc': acc,
@@ -217,9 +197,12 @@ def transfer_train(data_config: dict,
         'ref_label': ref_label,
         'query_raw_data': query_data,
         'query_label': query_label,
-        'pred': pred
+        'pred': pred,
+        'cpm_model': cpm_model,
+        'gcn_models': ref_gcn_models
     }
     return ret
+
 
 def show_result(ret):
     embedding_h = np.concatenate([ret['ref_h'], ret['query_h']], axis=0)
@@ -251,23 +234,25 @@ def show_result(ret):
         'Batch Mixing Entropy Mean': bme
     })
 
-    raw_data = np.concatenate([ret['ref_raw_data'], ret['query_raw_data']], axis=0)
-    # raw_data_pca = runPCA(raw_data)
-    #
-    raw_data_2d = runUMAP(raw_data)  # 对PCA之后的数据进行UMAP可视化
-    ref_len = ret['ref_raw_data'].shape[0]
-    #
-    h_data_2d = runUMAP(embedding_h)
-
-    np.save('result/raw_data_2d.npy', raw_data_2d)
-    np.save('result/h_data_2d.npy', h_data_2d)
-
     all_true_labels = np.concatenate([ret['ref_label'], ret['query_label']]).reshape(-1)
     all_pred_labels = np.concatenate([ret['ref_label'], ret['pred']]).reshape(-1)
+
+    if not parameter_config['model_exist']:
+        raw_data = np.concatenate([ret['ref_raw_data'], ret['query_raw_data']], axis=0)
+        # raw_data_pca = runPCA(raw_data)
+        #
+        raw_data_2d = runUMAP(raw_data)  # 对PCA之后的数据进行UMAP可视化
+        # ref_len = ret['ref_raw_data'].shape[0]
+        #
+        np.save('result/raw_data_2d.npy', raw_data_2d)
+        show_cluster(raw_data_2d, all_true_labels, 'reference-query raw true label')
+
+    h_data_2d = runUMAP(embedding_h)
+    np.save('result/h_data_2d.npy', h_data_2d)
     np.save('result/all_true_labels.npy', all_true_labels)
     np.save('result/all_pred_labels.npy', all_pred_labels)
 
-    show_cluster(raw_data_2d, all_true_labels, 'reference-query raw true label')
+
     show_cluster(h_data_2d, all_true_labels, 'reference-query h true label')
     show_cluster(h_data_2d, all_pred_labels, 'reference-query h pred label')
 
@@ -285,34 +270,47 @@ def show_result(ret):
     # show_cluster(raw_data_2d, np.concatenate([['Reference' for i in range(len(ret['ref_label']))], ['Query' for i in range(len(ret['query_label']))]])
     #              , 'Reference-Query Raw')
 
+def save_models(gcn_models, cpm_model):
+    for i, model in enumerate(gcn_models):
+        torch.save(model, 'result/gcn_model_'+ str(i) + '.pt')
+    torch.save(cpm_model, 'result/cpm_model.pt')
+
+
+
+def load_models():
+    gcn_models = []
+    for i in range(4):
+        model = torch.load('result/gcn_model_'+ str(i) + '.pt')
+        gcn_models.append(model)
+    cpm_model = torch.load('result/cpm_model.pt')
+
+    return gcn_models, cpm_model
 
 # 数据配置
 data_config = {
-    'data_path': 'C:\\Users\\WRX80\\Desktop\\kevislin\\data\\species\\GSE84133',
-    'ref_name': 'mouse',
-    'query_name': 'human',
+    'data_path': 'C:\\Users\\WRX80\\Desktop\\kevislin\\data\\species\\GSE84133\\data.h5',
+    'ref_name': 'GSE84133: mouse',
+    'query_name': 'GSE84133: human',
+    'query_key': 'query_1',
     'project': 'species',
     'class_num': 8,
-    'dataset_name':'GSE84133'
+    'dataset_name': 'GSE84133',
 }
-#{'gamma', 'alpha', 'endothelial', 'macrophage', 'ductal', 'delta', 'beta', 'quiescent_stellate'}
+# {'gamma', 'alpha', 'endothelial', 'macrophage', 'ductal', 'delta', 'beta', 'quiescent_stellate'}
 
-config = {
-    'epoch_GCN': 2000,  # Huang model 训练的epoch
-    'epoch_CPM_train': 3000,
-    'epoch_CPM_test': 5000,
-    'lsd_dim': 64,  # CPM_net latent space dimension
-    'GNN_lr': 0.001,
-    'CPM_lr': [0.001, 0.001, 0.001],  # CPM_ner中net和train_h,test_h的学习率
+parameter_config = {
     'ref_class_num': data_config['class_num'],  # Reference data的类别数
-    'query_class_num': data_config['class_num'],  # query data的类别数
+    'epoch_GCN': 1,  # Huang model 训练的epoch
+    'epoch_CPM_train': 1,
+    'epoch_CPM_test': 1,
+    'lsd_dim': 64,  # CPM_net latent space dimension
     'k': 2,  # 图构造的时候k_neighbor参数
     'middle_out': 2048,  # GCN中间层维数
     'w_classify': 10,  # classfication loss的权重
     'cen_weight': 0.5,
     'mask_rate': 0.3,
+    'model_exist': False,  # 如果事先已经有了模型,则为True
 }
-
 
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
@@ -320,48 +318,21 @@ print("Transfer across " + data_config['project'])
 print("Reference: " + data_config['ref_name'], "Query: " + data_config['query_name'])
 
 
-def main_process(data_config, config):
+def main_process():
     run = wandb.init(project="cell_classify_" + data_config['project'],
                      entity="kevislin",
-                     config={"config": config, "data_config": data_config},
-                     tags=[data_config['ref_name'] + '-' + data_config['query_name'], data_config['project'],
-                           str(data_config['class_num']) + ' class', data_config['dataset_name'], "Adjust: lsd_dim"],
+                     config={"config": parameter_config, "data_config": data_config},
+                     tags=[data_config['ref_name'] + '-' + data_config['query_name'], data_config['project']],
                      reinit=True)
-    ret = transfer_train(data_config, config)
+    ret = transfer_labels()
+    # 保存模型, 利用pickle
+    save_models(ret['gcn_models'], ret['cpm_model'])
+    # 查看结果
     show_result(ret)
+
     run.finish()
 
 
 # 测试epoch_CPM_train
-
-
-config['mask_rate'] = 0.05
-main_process(data_config, config)
-
-config['mask_rate'] = 0.1
-main_process(data_config, config)
-
-config['mask_rate'] = 0.15
-main_process(data_config, config)
-
-config['mask_rate'] = 0.2
-main_process(data_config, config)
-
-config['mask_rate'] = 0.25
-main_process(data_config, config)
-
-config['mask_rate'] = 0.3
-main_process(data_config, config)
-
-config['mask_rate'] = 0.4
-main_process(data_config, config)
-
-config['mask_rate'] = 0.5
-main_process(data_config, config)
-
-
+main_process()
 # main_process(data_config, config)
-
-
-
-
